Delete dist_metrics.py
When we had a brief two-day period of AMPL on COLAB installation issues, for Columbia class lecture continuity, we used a stand-alone version of COLAB notebook that included this and other AMPL python function chunks. Since the issues are fixed, we do not need them. I am deleting them from the branch.