Save best chromosome, sanity check chromosome, increase test duration
Added a mechanism that saves the best chromosome/solution across
all generations. I observed the best_fitness decrease a few times
and this mechanism will prevent a sub optimial solution.

Added a function that sanity checks a chromosome to make sure that
every task has a non-na label for every subset. This wasn't done
before and was crashing in the wasserstein fitness function.

Increased the number of generations and super scaffolds to give
the test a better chance of succeeding.